+++ src/main/java/com/speedit/inventorysystem/dto/stockmonitoring/ProductJsonDTO.java
@@ -3,10 +3,12 @@ package com.speedit.inventorysystem.dto.stockmonitoring;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 public class ProductJsonDTO {
     private int productId;
     private String productInfo;
-    private double price; // Use double for decimal format like 1.50
+    private BigDecimal price;
 }
